Fix the assignements/PA{2,4}/Makefile to include itself in the SRC variable.
The instruction in the assignment PDF below is incorrect. The Makefile
in assignments/PA{2,4} did not include itself in the SRC variable. This
caused the Makefile to not be copied to the workspace directory when the
workspace was created.
```
To get started, create a directory where you want to do the assignment and execute one of the following
commands in that directory. For the C++ version of the assignment, you should type

make -f /usr/class/cs143/assignments/PA4/Makefile
```